Moved Gyro over to navagation subsystem
