+++ src/main/java/frc/robot/subsystems/DriveSubsystem.java
@@ -4,12 +4,12 @@ import java.text.DecimalFormat;
 
 import com.ctre.phoenix.motorcontrol.ControlMode;
 import com.ctre.phoenix.motorcontrol.can.TalonSRX;
-import com.kauailabs.navx.frc.AHRS;
+// import com.kauailabs.navx.frc.AHRS;
 
-import edu.wpi.first.wpilibj.DriverStation;
+// import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.PIDController;
 import edu.wpi.first.wpilibj.PIDOutput;
-import edu.wpi.first.wpilibj.SPI;
+// import edu.wpi.first.wpilibj.SPI;
 import edu.wpi.first.wpilibj.command.Subsystem;
 import frc.lib.drive.ArcadeDrive;
 import frc.lib.drive.TankDrive;
@@ -20,7 +20,7 @@ import frc.robot.commands.SplitArcadeCommand;
  * Add your docs here.
  */
 public class DriveSubsystem extends Subsystem implements PIDOutput {
-	private AHRS gyro;
+//   private AHRS gyro;
   private PIDController pidController;
 
   public double PIDOutput;
@@ -159,18 +159,18 @@ public class DriveSubsystem extends Subsystem implements PIDOutput {
 		leftSlaveZero.setInverted(true);
 		leftSlaveOne.setInverted(true);
 
-		try {
-			gyro = new AHRS(SPI.Port.kMXP);
-		} catch(RuntimeException ex) {
-			DriverStation.reportError("Error instantiating navX MXP: " + ex.getMessage(), true);
-		}
+		// try {
+		// 	gyro = new AHRS(SPI.Port.kMXP);
+		// } catch(RuntimeException ex) {
+		// 	DriverStation.reportError("Error instantiating navX MXP: " + ex.getMessage(), true);
+		// }
 
-		gyro.zeroYaw();
+		// gyro.zeroYaw();
 
 		pidController = new PIDController(Robot.m_constraintsProperties.getkP(), 
 										  Robot.m_constraintsProperties.getkI(), 
 										  Robot.m_constraintsProperties.getkD(), 
-										  gyro, 
+										  Robot.m_navagationSubsystem.gyro, 
 										  this);
 		
 		pidController.disable();
@@ -248,21 +248,21 @@ public class DriveSubsystem extends Subsystem implements PIDOutput {
   }
 
   public boolean pidDone() {
-	  if(Math.abs(Math.abs(pidController.getSetpoint()) - Math.abs(gyro.getYaw())) < Robot.m_constraintsProperties.getPIDTolerance()) {
+	  if(Math.abs(Math.abs(pidController.getSetpoint()) - Math.abs(Robot.m_navagationSubsystem.gyro.getYaw())) < Robot.m_constraintsProperties.getPIDTolerance()) {
 		  return true;
 	  } else {
 		  return false;
 	  }
   }
 
-  public boolean resetYaw() {
-	  gyro.zeroYaw();
-	  return true;
-  }
+//   public boolean resetYaw() {
+// 	  gyro.zeroYaw();
+// 	  return true;
+//   }
 
-  public double getYaw() {
-	  return gyro.getYaw();
-  }
+//   public double getYaw() {
+// 	  return gyro.getYaw();
+//   }
   
   @Override
   public void initDefaultCommand() {
